Adds useToast custom hook
